\name{ymat}
\alias{ymat}
\docType{data}

\title{Response Variable Matrix 
}

\description{
This is a \eqn{26} by \eqn{36} matrix. Each column is a trajectory of Landsat signals corresponding to \eqn{26} consecutive years ranging from \eqn{1985} to \eqn{2010}. It will be used in some examples of this package. They are trajectories from \eqn{36} pixels in South Carolina.}

\usage{data("ymat")}

\format{
A \eqn{26} by \eqn{36} matrix. 
}

\source{US Forest Service}

\seealso{\code{\link{shape}}, \code{\link{plotshape}}}

\keyword{datasets}
